+++ cucumber-test/src/main/java/TestRotationLeft.java
@@ -1,16 +1,16 @@
 import static org.junit.Assert.*;
 
+
 import org.junit.Test;
 
 import decks.Deck;
 import model.tile.Robot;
-import subcards.Card;
-import subcards.RotateLeft;
-import subcards.RotateRight;
+import model.cards.*;
+
 
 public class TestRotationLeft {
 	
-	Deck deck = new Deck();
+	Deck deck = Deck.getInstance();
 	Robot robot = new Robot(0, 0);
 	RotateLeft rotateLeftCard = new RotateLeft(0);
