+++ cucumber-test/src/main/java/model/tile/Robot.java
@@ -1,12 +1,11 @@
 package model.tile;
-import model.subcards.*;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import decks.Deck;
 import decks.SubDeck;
+import model.cards.*;
 import utilities.Position;
